feat: Simplify MCP server configuration and response handling with optimized structures and error management
