+++ pkg/filter/mcp/mcpserver/filter.go
@@ -22,39 +22,18 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"strings"
+)
 
-	"github.com/apache/dubbo-go-pixiu/pkg/model"
-
-	"github.com/apache/dubbo-go-pixiu/pkg/common/extension/filter"
-	h "github.com/apache/dubbo-go-pixiu/pkg/context/http"
-	"github.com/apache/dubbo-go-pixiu/pkg/logger"
+import (
 	"github.com/creasty/defaults"
 	"github.com/mark3labs/mcp-go/mcp"
 )
 
-// MCP method constants using mcp-go library constants
-const (
-	methodInitialize            = string(mcp.MethodInitialize)
-	methodToolsList             = string(mcp.MethodToolsList)
-	methodToolsCall             = string(mcp.MethodToolsCall)
-	methodResourcesList         = string(mcp.MethodResourcesList)
-	methodResourcesRead         = string(mcp.MethodResourcesRead)
-	methodResourceTemplatesList = "resources/templates/list"
-	methodPromptsList           = string(mcp.MethodPromptsList)
-	methodPromptsGet            = string(mcp.MethodPromptsGet)
-	methodNotificationsInit     = "notifications/initialized"
-	methodPing                  = string(mcp.MethodPing)
-)
-
-// HTTP status code ranges
-const (
-	httpStatusClientErrorStart = 400
-)
-
-// MCP protocol constants
-const (
-	mcpProtocolVersion = "2024-11-05"
+import (
+	"github.com/apache/dubbo-go-pixiu/pkg/common/constant"
+	"github.com/apache/dubbo-go-pixiu/pkg/common/extension/filter"
+	h "github.com/apache/dubbo-go-pixiu/pkg/context/http"
+	"github.com/apache/dubbo-go-pixiu/pkg/logger"
 )
 
 // FilterFactory and MCPServerFilter types
@@ -116,10 +95,6 @@ func (f *FilterFactory) Apply() error {
 		logger.Debugf("[dubbo-go-pixiu] mcp server registered prompt '%s'", prompt.Name)
 	}
 
-	toolCount, resourceCount, templateCount, promptCount := f.registry.Count()
-	logger.Infof("[dubbo-go-pixiu] mcp server initialized successfully - tools:%d, resources:%d, templates:%d, prompts:%d",
-		toolCount, resourceCount, templateCount, promptCount)
-
 	return nil
 }
 
@@ -133,8 +108,8 @@ func (f *FilterFactory) PrepareFilterChain(ctx *h.HttpContext, chain filter.Filt
 	mcpFilter := &MCPServerFilter{
 		cfg:             f.cfg,
 		registry:        f.registry,
-		errorHandler:    NewErrorHandler(),
-		responseBuilder: NewResponseBuilder(),
+		errorHandler:    GetErrorHandler(),
+		responseBuilder: GetResponseBuilder(),
 	}
 	chain.AppendDecodeFilters(mcpFilter)
 	chain.AppendEncodeFilters(mcpFilter) // Add to Encode chain
@@ -174,7 +149,7 @@ func (f *MCPServerFilter) Decode(ctx *h.HttpContext) filter.FilterStatus {
 	// Handle terminal methods (methods that don't need forwarding to backend)
 	if f.isTerminalMethod(jsonrpcReq.Method) {
 		return f.handleTerminalMethod(mcpCtx, jsonrpcReq)
-	} else if jsonrpcReq.Method == methodToolsCall {
+	} else if jsonrpcReq.Method == string(mcp.MethodToolsCall) {
 		// Tool call will be processed in Encode stage (IsMCPToolCall() checks method)
 		return f.handleToolCall(mcpCtx, jsonrpcReq)
 	} else {
@@ -187,9 +162,9 @@ func (f *MCPServerFilter) Decode(ctx *h.HttpContext) filter.FilterStatus {
 // isTerminalMethod checks if it's a terminal method (methods that don't need forwarding to backend)
 func (f *MCPServerFilter) isTerminalMethod(method string) bool {
 	switch method {
-	case methodInitialize, methodToolsList, methodResourcesList, methodResourcesRead,
-		methodResourceTemplatesList, methodPromptsList, methodPromptsGet,
-		methodNotificationsInit, methodPing:
+	case string(mcp.MethodInitialize), string(mcp.MethodToolsList), string(mcp.MethodResourcesList), string(mcp.MethodResourcesRead),
+		"resources/templates/list", string(mcp.MethodPromptsList), string(mcp.MethodPromptsGet),
+		"notifications/initialized", string(mcp.MethodPing):
 		return true
 	default:
 		return false
@@ -199,23 +174,23 @@ func (f *MCPServerFilter) isTerminalMethod(method string) bool {
 // handleTerminalMethod handles terminal methods
 func (f *MCPServerFilter) handleTerminalMethod(ctx *MCPContext, req mcp.JSONRPCRequest) filter.FilterStatus {
 	switch req.Method {
-	case methodInitialize:
+	case string(mcp.MethodInitialize):
 		return f.handleInitialize(ctx, req)
-	case methodToolsList:
+	case string(mcp.MethodToolsList):
 		return f.handleToolsList(ctx, req)
-	case methodResourcesList:
+	case string(mcp.MethodResourcesList):
 		return f.handleResourcesList(ctx, req)
-	case methodResourcesRead:
+	case string(mcp.MethodResourcesRead):
 		return f.handleResourceRead(ctx, req)
-	case methodResourceTemplatesList:
+	case "resources/templates/list":
 		return f.handleResourceTemplatesList(ctx, req)
-	case methodPromptsList:
+	case string(mcp.MethodPromptsList):
 		return f.handlePromptsList(ctx, req)
-	case methodPromptsGet:
+	case string(mcp.MethodPromptsGet):
 		return f.handlePromptsGet(ctx, req)
-	case methodNotificationsInit:
+	case "notifications/initialized":
 		return f.handleNotificationsInitialized(ctx, req)
-	case methodPing:
+	case string(mcp.MethodPing):
 		return f.handlePing(ctx, req)
 	default:
 		logger.Warnf("[dubbo-go-pixiu] mcp server unsupported method: %s", req.Method)
@@ -263,134 +238,7 @@ func (f *MCPServerFilter) sendJSONResponse(ctx *MCPContext, response any) filter
 	logger.Infof("[dubbo-go-pixiu] mcp server response sent: %s (id: %v)", method, requestID)
 
 	// Critical: Clear Content-Length header to prevent mismatch errors
-	ctx.Writer.Header().Del("Content-Length")
+	ctx.Writer.Header().Del(constant.HeaderKeyContentLength)
 	ctx.SendLocalReply(http.StatusOK, responseBody)
 	return filter.Stop
-}
-
-// handleToolCall handles tool call requests by forwarding to backend
-func (f *MCPServerFilter) handleToolCall(ctx *MCPContext, req mcp.JSONRPCRequest) filter.FilterStatus {
-	// Parse tool call parameters
-	paramsBytes, err := json.Marshal(req.Params)
-	if err != nil {
-		logger.Errorf("[dubbo-go-pixiu] mcp server failed to marshal tool call params: %v", err)
-		return f.errorHandler.SendInternalError(ctx, req.ID, "invalid tool call parameters")
-	}
-
-	var params struct {
-		Name      string         `json:"name"`
-		Arguments map[string]any `json:"arguments,omitempty"`
-	}
-	if err := json.Unmarshal(paramsBytes, &params); err != nil {
-		logger.Errorf("[dubbo-go-pixiu] mcp server failed to parse tool call params: %v", err)
-		return f.errorHandler.SendInvalidParams(ctx, req.ID, "invalid tool call parameters")
-	}
-
-	// Find tool configuration
-	toolConfig, exists := f.registry.GetTool(params.Name)
-	if !exists {
-		logger.Warnf("[dubbo-go-pixiu] mcp server tool not found: %s", params.Name)
-		return f.errorHandler.SendToolCallError(ctx, req.ID, fmt.Sprintf("tool not found: %s", params.Name))
-	}
-
-	// Build backend request
-	err = f.buildBackendRequest(ctx, toolConfig, params.Arguments)
-	if err != nil {
-		logger.Errorf("[dubbo-go-pixiu] mcp server failed to build backend request: %v", err)
-		return f.errorHandler.SendToolCallError(ctx, req.ID, "failed to build backend request")
-	}
-
-	// Set cluster information for routing
-	if ctx.Params == nil {
-		ctx.Params = make(map[string]any)
-	}
-
-	logger.Infof("[dubbo-go-pixiu] mcp server forwarding tool call: %s -> %s %s (cluster: %s)",
-		params.Name, toolConfig.Request.Method, ctx.Request.URL.Path, toolConfig.Cluster)
-
-	// Store MCP data for Encode stage processing
-	ctx.StoreMCPDataInParams()
-
-	ctx.Route = &model.RouteAction{
-		Cluster: toolConfig.Cluster,
-	}
-
-	// Continue to next filter for backend forwarding
-	return filter.Continue
-}
-
-// buildBackendRequest builds the complete backend request including path, body, and headers
-func (f *MCPServerFilter) buildBackendRequest(ctx *MCPContext, toolConfig ToolConfig, arguments map[string]any) error {
-	// Set HTTP method
-	ctx.Request.Method = toolConfig.Request.Method
-
-	// Build request path and body based on argument locations
-	path := toolConfig.Request.Path
-	bodyParams := make(map[string]any)
-	queryParams := make(map[string]string)
-
-	// Process arguments based on their location (path, query, body)
-	if arguments != nil {
-		for argName, argValue := range arguments {
-			// Find argument configuration
-			var argConfig *ArgConfig
-			for _, arg := range toolConfig.Args {
-				if arg.Name == argName {
-					argConfig = &arg
-					break
-				}
-			}
-
-			if argConfig == nil {
-				continue // Skip unknown arguments
-			}
-
-			switch argConfig.In {
-			case "path":
-				// Replace path parameters
-				placeholder := fmt.Sprintf("{%s}", argName)
-				replacement := fmt.Sprintf("%v", argValue)
-				path = strings.ReplaceAll(path, placeholder, replacement)
-
-			case "query":
-				// Add to query parameters
-				queryParams[argName] = fmt.Sprintf("%v", argValue)
-
-			case "body":
-				// Add to request body
-				bodyParams[argName] = argValue
-			}
-		}
-	}
-
-	// Set the request path
-	ctx.Request.URL.Path = path
-
-	// Add query parameters
-	if len(queryParams) > 0 {
-		query := ctx.Request.URL.Query()
-		for key, value := range queryParams {
-			query.Set(key, value)
-		}
-		ctx.Request.URL.RawQuery = query.Encode()
-	}
-
-	// Build request body for POST/PUT requests
-	if len(bodyParams) > 0 && (toolConfig.Request.Method == "POST" || toolConfig.Request.Method == "PUT") {
-		bodyJSON, err := json.Marshal(bodyParams)
-		if err != nil {
-			return fmt.Errorf("failed to marshal request body: %v", err)
-		}
-
-		// Set request body
-		ctx.Request.Body = io.NopCloser(strings.NewReader(string(bodyJSON)))
-		ctx.Request.ContentLength = int64(len(bodyJSON))
-
-		// Set Content-Type header
-		ctx.Request.Header.Set("Content-Type", "application/json")
-
-		logger.Debugf("[dubbo-go-pixiu] mcp server built request body: %s", string(bodyJSON))
-	}
-
-	return nil
 }
