+++ pkg/filter/mcp/mcpserver/response.go
@@ -18,17 +18,36 @@
 package mcpserver
 
 import (
+	"encoding/json"
 	"fmt"
+	"net/http"
+	"sync"
+)
 
+import (
 	"github.com/mark3labs/mcp-go/mcp"
+)
+
+import (
+	"github.com/apache/dubbo-go-pixiu/pkg/common/constant"
+	"github.com/apache/dubbo-go-pixiu/pkg/common/extension/filter"
+	"github.com/apache/dubbo-go-pixiu/pkg/logger"
 )
 
 // ResponseBuilder provides methods to create standardized MCP responses
 type ResponseBuilder struct{}
 
-// NewResponseBuilder creates a new response builder
-func NewResponseBuilder() *ResponseBuilder {
-	return &ResponseBuilder{}
+var (
+	responseBuilderInstance *ResponseBuilder
+	responseBuilderOnce     sync.Once
+)
+
+// GetResponseBuilder returns the singleton ResponseBuilder instance
+func GetResponseBuilder() *ResponseBuilder {
+	responseBuilderOnce.Do(func() {
+		responseBuilderInstance = &ResponseBuilder{}
+	})
+	return responseBuilderInstance
 }
 
 // Success creates a successful JSON-RPC response
@@ -71,4 +90,63 @@ func (rb *ResponseBuilder) ToolCallError(id any, message string) mcp.JSONRPCResp
 	}
 
 	return rb.Success(id, result)
+}
+
+// ErrorHandler provides centralized error handling for MCP responses
+type ErrorHandler struct {
+	responseBuilder *ResponseBuilder
+}
+
+var (
+	errorHandlerInstance *ErrorHandler
+	errorHandlerOnce     sync.Once
+)
+
+// GetErrorHandler returns the singleton ErrorHandler instance
+func GetErrorHandler() *ErrorHandler {
+	errorHandlerOnce.Do(func() {
+		errorHandlerInstance = &ErrorHandler{
+			responseBuilder: GetResponseBuilder(),
+		}
+	})
+	return errorHandlerInstance
+}
+
+// SendInternalError sends an internal server error response
+func (eh *ErrorHandler) SendInternalError(ctx *MCPContext, id any, message string) filter.FilterStatus {
+	response := eh.responseBuilder.Error(id, mcp.INTERNAL_ERROR, message)
+	return eh.sendResponse(ctx, response)
+}
+
+// SendMethodNotFound sends a method not found error response
+func (eh *ErrorHandler) SendMethodNotFound(ctx *MCPContext, id any) filter.FilterStatus {
+	response := eh.responseBuilder.Error(id, mcp.METHOD_NOT_FOUND, "Method not found")
+	return eh.sendResponse(ctx, response)
+}
+
+// SendInvalidParams sends an invalid parameters error response
+func (eh *ErrorHandler) SendInvalidParams(ctx *MCPContext, id any, message string) filter.FilterStatus {
+	response := eh.responseBuilder.Error(id, mcp.INVALID_PARAMS, fmt.Sprintf("Invalid params: %s", message))
+	return eh.sendResponse(ctx, response)
+}
+
+// SendToolCallError sends a tool call error response
+func (eh *ErrorHandler) SendToolCallError(ctx *MCPContext, id any, message string) filter.FilterStatus {
+	response := eh.responseBuilder.ToolCallError(id, message)
+	return eh.sendResponse(ctx, response)
+}
+
+// sendResponse sends any response and handles Content-Length cleanup
+func (eh *ErrorHandler) sendResponse(ctx *MCPContext, response any) filter.FilterStatus {
+	responseBody, err := json.Marshal(response)
+	if err != nil {
+		logger.Errorf("[dubbo-go-pixiu] mcp server failed to marshal response: %v", err)
+		ctx.SendLocalReply(http.StatusInternalServerError, []byte("internal server error"))
+		return filter.Stop
+	}
+
+	// Critical: Clear Content-Length header to prevent mismatch errors
+	ctx.Writer.Header().Del(constant.HeaderKeyContentLength)
+	ctx.SendLocalReply(http.StatusOK, responseBody)
+	return filter.Stop
 }
